Move shrink wrap from input bar to top right
