+++ front/components/assistant/conversation/input_bar/InputBarContainer.tsx
@@ -17,18 +17,13 @@ import { useSendNotification } from "@app/hooks/useNotification";
 import { useVoiceTranscriberService } from "@app/hooks/useVoiceTranscriberService";
 import { getMcpServerViewDisplayName } from "@app/lib/actions/mcp_helper";
 import type { MCPServerViewType } from "@app/lib/api/mcp";
-import { useFeatureFlags } from "@app/lib/auth/AuthContext";
 import type { NodeCandidate, UrlCandidate } from "@app/lib/connectors";
 import { isNodeCandidate } from "@app/lib/connectors";
-import { useAppRouter } from "@app/lib/platform";
 import { getSkillIcon } from "@app/lib/skill";
 import { useSpaces, useSpacesSearch } from "@app/lib/swr/spaces";
 import { useIsMobile } from "@app/lib/swr/useIsMobile";
 import { classNames } from "@app/lib/utils";
-import {
-  getAgentBuilderRoute,
-  getManageSkillsRoute,
-} from "@app/lib/utils/router";
+import { getManageSkillsRoute } from "@app/lib/utils/router";
 import type { LightAgentConfigurationType } from "@app/types/assistant/agent";
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
 import type {
@@ -48,7 +43,6 @@ import {
   ArrowUpIcon,
   Button,
   Chip,
-  Cog6ToothIcon,
   cn,
   TextIcon,
   Toolbar,
@@ -134,13 +128,6 @@ const InputBarContainer = ({
   user,
 }: InputBarContainerProps) => {
   const isMobile = useIsMobile();
-  const router = useAppRouter();
-  const { hasFeature } = useFeatureFlags();
-  const canTurnIntoAgent =
-    actions.includes("turn-into-agent") &&
-    hasFeature("agent_builder_shrink_wrap") &&
-    conversation !== undefined &&
-    isBuilder(owner);
 
   const [nodeOrUrlCandidate, setNodeOrUrlCandidate] = useState<
     UrlCandidate | NodeCandidate | null
@@ -807,23 +794,6 @@ const InputBarContainer = ({
                       showFooterButtons={actions.includes(
                         "agents-list-with-actions"
                       )}
-                      disabled={disableInput}
-                    />
-                  )}
-                  {canTurnIntoAgent && (
-                    <Button
-                      variant="ghost-secondary"
-                      icon={Cog6ToothIcon}
-                      size={buttonSize}
-                      tooltip="Turn into agent"
-                      onClick={() => {
-                        const route = getAgentBuilderRoute(
-                          owner.sId,
-                          "new",
-                          `conversationId=${conversation.sId}`
-                        );
-                        void router.push(route);
-                      }}
                       disabled={disableInput}
                     />
                   )}
